==> 1. APOE Status.ipynb <==
import os
import hail as hl
hl.init(default_reference='GRCh38', idempotent=True)

from datetime import datetime
import os
import pandas as pd
import gzip
import subprocess

start = datetime.now()

bucket = os.getenv('WORKSPACE_BUCKET')
bucket

import hail as hl
from hail.plot import show
from bokeh.plotting import output_file, save
import bokeh.io
from bokeh.io import *
from bokeh.resources import INLINE
bokeh.io.output_notebook(INLINE) 
%matplotlib inline

# Set Hail storage path
mt_wgs_path = os.getenv("WGS_HAIL_STORAGE_PATH")
mt_wgs_path

# First Allele

mt = hl.read_matrix_table(mt_wgs_path)
mt = mt.drop("AD")

# Interval for 2 alelles. - Tried filtering by rsid but needed to annotate first.
test_intervals = ['chr19:44908684-44908685']

# Filter
mt = hl.filter_intervals(
    mt,
    [hl.parse_locus_interval(x,)
     for x in test_intervals])

# Split multi hits
mt = hl.split_multi_hts(mt)

# Set entries table
entries_table = mt.entries()

# Export to bucket as csv so code above only has to be run once!
entries_table.export(f'{bucket}/APOE_file_44908684.csv', delimiter=',')

# Text export
test = pd.read_csv(f'{bucket}/APOE_file_44908684.csv', on_bad_lines='skip')

# This snippet assumes you run setup first

# This code saves your dataframe into a csv file in a "data" folder in Google Bucket

# Replace df with THE NAME OF YOUR DATAFRAME
my_dataframe = test

# Replace 'test.csv' with THE NAME of the file you're going to store in the bucket (don't delete the quotation marks)
destination_filename = 'APOE_file_44908684.csv'

########################################################################
##
################# DON'T CHANGE FROM HERE ###############################
##
########################################################################

# save dataframe in a csv file in the same workspace as the notebook
my_dataframe.to_csv(destination_filename, index=True)

# get the bucket name
my_bucket = os.getenv('WORKSPACE_BUCKET')

# copy csv file to the bucket
args = ["gsutil", "cp", f"./{destination_filename}", f"{my_bucket}/data/"]
output = subprocess.run(args, capture_output=True)

# print output from gsutil
output.stderr


# Second Allele

mt = hl.read_matrix_table(mt_wgs_path)
mt = mt.drop("AD")

test_intervals = ['chr19:44908822-44908823']

mt = hl.filter_intervals(
    mt,
    [hl.parse_locus_interval(x,)
     for x in test_intervals])

mt = hl.split_multi_hts(mt)

entries_table = mt.entries()

entries_table.export(f'{bucket}/APOE_file_44908821.csv', delimiter=',')

test = pd.read_csv(f'{bucket}/APOE_file_44908821.csv', on_bad_lines='skip')

# This snippet assumes you run setup first

# This code saves your dataframe into a csv file in a "data" folder in Google Bucket

# Replace df with THE NAME OF YOUR DATAFRAME
my_dataframe = test

# Replace 'test.csv' with THE NAME of the file you're going to store in the bucket (don't delete the quotation marks)
destination_filename = 'APOE_file_44908821.csv'

########################################################################
##
################# DON'T CHANGE FROM HERE ###############################
##
########################################################################

# save dataframe in a csv file in the same workspace as the notebook
my_dataframe.to_csv(destination_filename, index=True)

# get the bucket name
my_bucket = os.getenv('WORKSPACE_BUCKET')

# copy csv file to the bucket
args = ["gsutil", "cp", f"./{destination_filename}", f"{my_bucket}/data/"]
output = subprocess.run(args, capture_output=True)

# print output from gsutil
output.stderr
